fix: 머지 후 dDay 타입 불일치 수정 (long → int)
TaxCalendarService.toResponse()에서 long → int 캐스팅,
TaxCalendarScheduler.ALERT_DAYS를 Set<Integer>로 변경.

+++ BE/src/main/java/com/ssafy/tax7i/taxcalendar/scheduler/TaxCalendarScheduler.java
@@ -19,7 +19,7 @@ import java.util.Set;
 @RequiredArgsConstructor
 public class TaxCalendarScheduler {
 
-    private static final Set<Long> ALERT_DAYS = Set.of(0L, 3L, 7L);
+    private static final Set<Integer> ALERT_DAYS = Set.of(0, 3, 7);
 
     private final TaxCalendarService taxCalendarService;
     private final FcmService fcmService;
